fix: complete index page
